move InsertSetAddGeneratedKeyToken to rewrite package
